% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plogF.r
\name{plogF}
\alias{plogF}
\title{Probability of Log Distribution}
\usage{
plogF(x)
}
\arguments{
\item{x}{Real number where -inf < x < inf.
CDF of X is F(x) = 1 - (1+5e^(-x))^(-0.1), -inf < x < inf.
In R, these limits determined by data limitations. Thus we
get the following piecewise function,
```
plogF(x) = (1 - (1 + 5 * exp(x))^(-0.2))
```}
}
\value{
CDF of X.
}
\description{
Computes CDF of X given the following:
X is a continuous random variable with PDF f(x) = e^x / (1+5e^x)^1.2 for -inf < x < inf.
X has the support space (-inf, inf).
The PDF of X is a member of the logF-family of diustributions which is often used in the
modeling of the log of lifetime data.
See Example 1.7.3 on page 52 of the book.
}
\examples{
cdf <- plogF(1)
cdf <- plogF(-1)
cdf <- plogF(3.14159)
# cdf <- plogF(real_number_variable_within_specified_limits)
}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to Mathematical Statistics, 8th Ed. Boston: Pearson
}
